Site admin now gets a warning when trying to delete a last group admin. And the user will not be deleted

+++ src/main/java/nl/miwgroningen/se6/heartcoded/CaTo/controller/UserController.java
@@ -2,6 +2,7 @@ package nl.miwgroningen.se6.heartcoded.CaTo.controller;
 
 import nl.miwgroningen.se6.heartcoded.CaTo.dto.UserDTO;
 import nl.miwgroningen.se6.heartcoded.CaTo.dto.UserRegistrationDTO;
+import nl.miwgroningen.se6.heartcoded.CaTo.service.MemberService;
 import nl.miwgroningen.se6.heartcoded.CaTo.service.TaskListService;
 import nl.miwgroningen.se6.heartcoded.CaTo.service.UserService;
 import org.springframework.stereotype.Controller;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 /**
  * @author Shalena Omapersad <shalenao@hotmail.com>
@@ -22,10 +24,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 @Controller
 public class UserController {
 
-    private UserService userService;
+    private static final String USER_IS_THE_LAST_ADMIN_IN_A_GROUP = "User is the last admin in a group";
 
-    public UserController(UserService userService) {
+    private UserService userService;
+    private MemberService memberService;
+
+    public UserController(UserService userService, MemberService memberService) {
         this.userService = userService;
+        this.memberService = memberService;
     }
 
     @GetMapping("/registration")
@@ -36,9 +42,13 @@ public class UserController {
     }
 
     @GetMapping("/users/delete/{userId}")
-    protected String deleteUser(@PathVariable("userId") Integer userId) {
+    protected String deleteUser(@PathVariable("userId") Integer userId, RedirectAttributes redirectAttributes) {
         if (!userService.currentUserIsSiteAdmin()) {
             return "redirect:/403";
+        }
+        if (memberService.userIsLastGroupAdminInAnyGroup(userId)) {
+            redirectAttributes.addAttribute("error", USER_IS_THE_LAST_ADMIN_IN_A_GROUP);
+            return "redirect:/siteAdminDashboard";
         }
         userService.deleteUserById(userId);
         return "redirect:/siteAdminDashboard";
